@@ -20,10 +20,10 @@ describe("The 'hlen' method", function () {
 
             it('reports the count of keys', function (done) {
                 var hash = 'test hash';
-                var field1 = new Buffer('0123456789');
-                var value1 = new Buffer('abcdefghij');
-                var field2 = new Buffer(0);
-                var value2 = new Buffer(0);
+                var field1 = Buffer.from('0123456789');
+                var value1 = Buffer.from('abcdefghij');
+                var field2 = Buffer.alloc(0);
+                var value2 = Buffer.alloc(0);
 
                 client.HSET(hash, field1, value1, helper.isNumber(1));
                 client.HSET(hash, field2, value2, helper.isNumber(1));
