@@ -4,7 +4,7 @@ var utils = require('./utils');
 var debug = require('./debug');
 var Multi = require('./multi');
 var Command = require('./command');
-var no_password_is_set = /no password is set/;
+var no_password_is_set = /no password is set|called without any password configured/;
 var loading = /LOADING/;
 var RedisClient = require('../').RedisClient;
 
